Streamlined build process
