+++ .gitignore
@@ -1,3 +1,5 @@
 .vscode/
 build/
 testing/
+bochsrc.bxrc
+bx_enh_dbg.ini
